componentize and add search to homepage
