+++ src/pages/index.js
@@ -1,67 +1,32 @@
 import React from "react";
 import g from "glamorous";
 import Link from "gatsby-link";
+import ClassTable from "../components/ClassTable";
 
 import { rhythm } from "../utils/typography";
 
-export default ({ data }) => {
-  return (
-    
-    <table>
-      <thead>
-        <tr>
-          <th>Name</th>
-          <th>District</th>
-          <th>Days</th>
-          <th>Start Date</th>
-          <th>End Date</th>
-          <th>Time</th>
-          <th>Grades</th>
-        </tr>
-      </thead>
-      <tbody>
-        {data.allCommunityEducationClasses.edges.map(({ node }) =>
-          node.fields ? <tr key={node.id}>
-            <td>
-              <Link to={node.fields.slug} >
-                {node.fields.className}
-              </Link>
-            </td>
-            <td>
-              <Link to={node.fields.slug} >
-                {node.fields.district}
-              </Link>
-            </td>
-            <td>
-              <Link to={node.fields.slug} >
-                {node.fields.days}
-              </Link>
-            </td>
-            <td>
-              <Link to={node.fields.slug} >
-                {node.fields.startdate}
-              </Link>
-            </td>
-            <td>
-              <Link to={node.fields.slug} >
-                {node.fields.enddate}
-              </Link>
-            </td>
-            <td>
-              <Link to={node.fields.slug} >
-                {node.fields.time}
-              </Link>
-            </td>
-            <td>
-              <Link to={node.fields.slug} >
-                {node.fields.grades}
-              </Link>
-            </td>
-          </tr> : ''
-        )}
-      </tbody>
-    </table>
-  )
+export default class Index extends React.Component {
+  state = {
+    searchText: '',
+  }
+
+  updateSearchText = (event) => {
+    this.setState({
+      searchText: event.target.value,
+    });
+  }
+
+  render() {
+    return (
+      <div>
+        <input type="text" onChange={this.updateSearchText} value={this.state.searchText} />
+        <ClassTable
+          districtClasses={this.props.data.allCommunityEducationClasses.edges}
+          searchText={this.state.searchText}
+        />
+      </div>
+    )
+  }
 }
 
 export const query = graphql`
@@ -70,6 +35,7 @@ export const query = graphql`
       totalCount
       edges {
         node {
+          id
           fields {
             className
             days
